Fix prepareJSONNode for array normalization
When the content of a JSON node is an array, a similar error is thrown:

```
Error: Names must match /^[_a-zA-Z][_a-zA-Z0-9]*$/ but "0___image" does not.
```

This is caused by the spread operator transforming the index of the array to numbered keys of the object, resulting in an error because the name doesn't comply the GraphQL naming rules.

In order not to break the compatibility I keep spreading the object if the content of the JSON is a plain object, but I nest it under `items` if it's an array. This could be changed and made more consistent in the next major update.